// mat4331-basetalk-crawler/src/common/types/pitch-info.type.ts
export interface PitchInfo {
  IP: number;
  TBF: number;
  H: number;
  R: number;
  ER: number;
  BB: number;
  HBP: number;
  K: number;
  HR: number;
  GO: number;
  FO: number;
  NP: number;
  S: number;
  IR: number;
  IS: number;
  GSC: number;
  ERA: number;
  WHIP: number;
  LI: number;
  WPA: number;
  RE24: number;
}